# web-portal/backend/openapi.yaml
openapi: 3.0.0
info:
  title: CareConnect Backend API
  description: Backend service for the Timesheet frontend.
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Development server
paths:
  /entries:
    get:
      summary: Get timesheet
      responses:
        '200':
          description: A timesheet
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GetWorklogItem'
        '400':
          description: Missing or invalid parameters
        '401':
          description: Authentication failed
        '500':
          description: Internal server error
  /entry:
    post:
      summary: Post timesheet entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostWorklogItem'
      responses:
        '201':
          description: A timesheet
        '400':
          description: Missing or invalid parameters
        '401':
          description: Authentication failed
        '500':
          description: Internal server error
components:
  schemas:
    GetWorklogItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: string
          example: "123ABC"
        date:
          type: string
          format: date
          example: "2024-04-01"
        project:
          type: string
          example: "Project A"
        worklog:
          type: string
          example: "Worklog for Project A"
        duration:
          type: integer
          example: 3
    PostWorklogItem:
      type: object
      properties:
        user_id:
          type: string
          example: "123ABC"
        date:
          type: string
          format: date
          example: "2024-04-01"
        project:
          type: string
          example: "Project A"
        worklog:
          type: string
          example: "Worklog for Project A"
        duration:
          type: integer
          example: 3